E2ETests: test raw configuration
In order to test that the raw configuration works we manually generate a
neighbor configuration and we append to the rendered one.

To test that the merging of multiple configurations work, we also split
the configuration in two, one to define the neighbor address and the
other to enable the ip family.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>