Manifests: Add separate webhook pod
In contrast to MetalLB, we don't have a controller pod and we can't have the
webhooks ran on each daemon as the cert rotation clashes. We add a separate
pod to act as a webhook-server only, using the same bin as the daemon.
This might be changed if/when we use something else to handle the certs.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>