Add option to enable the rotator RestartOnSecretRefresh
When the webhook pod is redeployed multiple times in succession sometimes the requests for
the webhook hang on `tls: failed to verify certificate` and it is not solved until
the pod is restarted manually. While the reason the flag solves the issue is not
totally clear, introducing the flag for the operator to consume is reasonable.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>