Remove the port from the peer metrics label
For each VRF we can have only one neighbor for a given address, which
makes the port irrelevant.

Additionally, now that frr-k8s can accept connections too, the port can
either be the one used on our side to establish the connection, or
the source port in case we are accepting a bgp connection.
This makes the metric labels non predictable and hard to monitor.

Because of this, here we remove the port from the label.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>