Add zebra -K graceful_restart argument
https://docs.frrouting.org/en/latest/zebra.html#cmdoption-zebra-K

With that argument `-K 120`, when zebra stops it does not clean the
routes that has install on the host. When zebra starts up with, it will
wait 120 seconds before removing (sweeping) any routes that it
previously created but are no longer valid.

The value 120 is used to match the default `bgp gracefule-restart
restart-time` that we advertise to the peer which is also 120seconds.
https://docs.frrouting.org/en/latest/bgp.html#clicmd-bgp-graceful-restart-restart-time-0-4095

This is needed to complete the BGP graceful restart feature when the
local endpoints depend on local routes that are learned from the BGP
peer.

Downside is that if we simply remove FRR (delete deployment or change
speaker label toleration) those routes will remain on the host until
reboot.

Signed-off-by: karampok <karampok@gmail.com>