Add zebra -K --graceful_restart argument
https://docs.frrouting.org/en/latest/zebra.html#cmdoption-zebra-K

With that argument `-K 120`, when zebra stops it does not clean the
routes that has install on the host. When zebra starts up with, it will
wait 120 seconds before removing (sweeping) any routes that it
previously created but are no longer valid.

This is needed to complete the BGP graceful restart feature when the
local endpoints depend on local routes that are learned from the BGP
peer.

Downside is that if we simply remove FRR (delete deployment or change
speaker label toleration) those routes will remain on the host until
reboot.

Signed-off-by: karampok <karampok@gmail.com>