Webhook pod: report ready / live when the webhook endpoint is ready
Instead of relying on the metrics port, we move to test the webhook
endpoint. The webhook will declare itself ready only when the
cert controller did its part and the certificate is generated correctly.

This allows the crs not to fail on creation if we apply them after
checking all pods are ready (ie right after a make deploy).

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>