E2E: Fix BGPSessionState CI flake
There is currently a bug on FRR 9.x that BFD profiles
are not removed properly when a config that has BFD
is replaced by one that does not, causing these tests
to fail when the Clean isn't applied before the
tests themselves (because of the debouncer).
We put a temporary fix for that, to be removed
once we move to FRR 10.x or the bug is backported to
FRR 9.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>