FRR: provide a go api more similar to the frr one
Having an incoming api more aligned to what FRR needs has the great
benefit of simplify the template.

The template for generating the frr configuration was too smart, and
hard to follow / reason about. Having an incoming config more aligned to
what frr requires removes the burden of having extra logic in the
template and moves it to the caller side, which must provide the valid
configuration.

The advantage here is that we do that effort in Go, which is easier to
follow and to cover with tests.

The FRR-k8s api were thought with the original (metallb) frr api in
mind.
By changing the prospective, the frr package  now accepts a configuration that is
more similar to what frr expects. In order to do the conversion, an
extra effort is now required by the controller, which need to translate
the user facing api to the internal one.
Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>