Controller: align the conversion / merging to the new FRR API
The FRR-k8s api were thought with the original (metallb) frr api in
mind.
By changing the prospective, the frr package  now accepts a configuration that is
more similar to what frr expects. In order to do the conversion, an
extra effort is now required by the controller, which need to translate
the user facing api to the internal one.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>