E2E: Add a simple session status test that does not rely on polling
We add simple coverage for the session status resource that does
not rely on the poll-interval, so that we can verify the feature
on clusters where we don't have control over it.
More specifically, the existing (more extensive) E2E relies on
the resource being updated by what we poll from FRR, e.g reflecting
when the BGP session establishes, and in order to have it run relatively
fast the poll-interval of the status exporter needs to be low so that
the statuses are updated quickly (as we hardcode in CI to 5s).

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>